+++ packages/components/fields/radio-field/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './radio-field';
+
+export { default as version } from './version';
