+++ packages/components/fields/radio-field/src/radio-field.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { RadioField, RadioInput } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const value = 'apple';
