--- src/components/fields/radio-field/index.js
@@ -1,1 +1,0 @@
-export { default } from './radio-field';
